pyproject.toml
[project]
name = "datathon2026-challenge-harness"
version = "0.1.0"
description = "Starter FastAPI harness for the Datathon 2026 listings challenge."
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.0,<2.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "fastapi[standard]",
    "httpx>=0.28.0,<1.0.0",
    "mcp>=1.25.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    "pytest",
    "pydantic-ai-slim[bedrock]>=1.84.1",
    "python-dotenv>=1.2.2",
    "ipykernel>=7.2.0",
    "hydra-core>=1.3.2",
    "requests",
    "opensearch-py==3.1.0",
    "tqdm>=4.0.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "apps_sdk"]

[tool.pytest.ini_options]
pythonpath = ["."]
